新たなカットシーンの追加

+++ project/Source/Weapon/CharaWeapon.cpp
@@ -89,7 +89,7 @@ void CharaWeapon::ImguiDraw()
 	}
 }
 
-void CharaWeapon::CreateSwordEffect(const VECTOR3& _nPos, const VECTOR3& _fPos, float _r, float _g, float _b, float _a, float index, float _time)
+void CharaWeapon::CreateSwordEffect(const VECTOR3& _nPos, const VECTOR3& _fPos, float _r, float _g, float _b, float _a, int index, float _time)
 {
 	
 	SwordEffect* sE = obj->Component()->GetComponent<SwordEffect>();
